tests/wpt/web-platform-tests/content-security-policy/unsafe-hashes/javascript_src_denied_missing_unsafe_hashes-window_location.html

<!DOCTYPE HTML>
<html>

<head>
    <!--
    'sha256-IIiAJ8UuliU8o1qAv6CV4P3R8DeTf/v3MrsCwXW171Y=' ==> 'javascript:opener.postMessage('pass', '*')'
    -->
    <script src='/resources/testharness.js' nonce='abc'></script>
    <script src='/resources/testharnessreport.js' nonce='abc'></script>
</head>

<body>
    <div id='log'></div>
    <script nonce='abc'>
        var t1 = async_test("Test that the javascript: src is not allowed to run");

        window.onmessage = t1.step_func_done(function(e) {
          assert_equals(e.data, "fail");
        });

        window.open('support/child_window_location_navigate.sub.html' +
              '?csp=' + encodeURI("script-src 'nonce-abc' 'sha256-IIiAJ8UuliU8o1qAv6CV4P3R8DeTf/v3MrsCwXW171Y='") +
              '&url=' + encodeURI("javascript:opener.postMessage('pass', '*')"));
    </script>
</body>

</html>
